Feature/grades (#183)
* update subject handling to use SubjectClassDetail and change grade type to Double

* add percentage grade sum calculations and validation for exercises

+++ codefm-domain/src/main/java/org/web/codefm/domain/entity/teachernotebook/ExerciseStudentGrade.java
@@ -13,7 +13,7 @@ public class ExerciseStudentGrade {
     private Integer id;
     private Integer studentId;
     private Integer exerciseId;
-    private Integer grade;
+    private Double grade;
     private String description;
     private LocalDate deletionDate;
     private String studentName;
